Fixing bugs with the argument inspector
switch statement was borked becuase it wasn't breaking so it was always
displaying as string arguments.

Fixed a bug where bad input on a float argument would stick when
switching functions being inspected if the other function also had a
float argument at the same index. The key for the array now uses the id
for the function call as well.