refactoring
Moved the function state to be under the statement state. That way
adding a new function call can be handled in one spot, rather than
having to make it two separate actions.

Added a default action type to quite typescript errors interfacing with
the generic actions in the redux api, while still allowing typesafety
inside the reducer's switch statements.

Setup the DevTools window and added the react and redux extensions.